show bridge type badge for ictt
